requirements.txt
fusepy
upsilon-py
# Upsilon forgot to add it as a dependency
rich